+++ src/integrators/mlt.h
@@ -62,7 +62,7 @@ class MLTSampler : public Sampler {
           streamCount(streamCount) {}
     Float Get1D();
     Point2f Get2D();
-    std::unique_ptr<Sampler> Clone(int seed);
+    std::unique_ptr<Sampler> Clone(uint64_t seed);
     void StartIteration();
     void Accept();
     void Reject();
