+++ src/samplers/zerotwosequence.h
@@ -49,7 +49,7 @@ class ZeroTwoSequenceSampler : public PixelSampler {
     // ZeroTwoSequenceSampler Public Methods
     ZeroTwoSequenceSampler(int64_t samplesPerPixel, int nSampledDimensions = 4);
     void StartPixel(const Point2i &);
-    std::unique_ptr<Sampler> Clone(int seed);
+    std::unique_ptr<Sampler> Clone(uint64_t seed);
     int RoundCount(int count) const { return RoundUpPow2(count); }
 };
